Changed the accepted value in pagination  from integer to long

+++ src/main/java/com/kolos/bookstore/service/util/PageUtil.java
@@ -4,8 +4,8 @@ import com.kolos.bookstore.service.dto.PageableDto;
 
 public class PageUtil {
 
-    public static int getTotalPages(PageableDto pageableDto, int count) {
-        int pages = count / pageableDto.getPageSize();
+    public static long getTotalPages(PageableDto pageableDto, long count) {
+        long pages = count / pageableDto.getPageSize();
         if (count % pageableDto.getPageSize() != 0) {
             pages ++;
         }
